refactor: GameSign, MoneyBetter, TurnProgress 함수형 인터페이스 패키지 이동
