+++ src/main/java/blackjack/model/game/MoneyBetter.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.model.game;
 
 @FunctionalInterface
 public interface MoneyBetter {
