+++ src/main/java/blackjack/model/game/TurnProgress.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.model.game;
 
 import blackjack.model.participant.Participant;
